==> 0_DataDownloader_PY.ipynb <==
import pandas as pd
import os
import mosqlient
import requests

base_dir = os.getcwd()
raw_data_dir = os.path.join(base_dir, "data_sprint_2025")

geo_df = pd.read_csv(os.path.join(raw_data_dir, "map_regional_health.csv"))

uf_list = geo_df["uf"].unique()

start_date = "2025-01-01"
end_date = "2025-07-05"
api_key = "add_api_key"

builder = []

for curr_uf in uf_list: 
    curr_geo_df = geo_df[geo_df["uf"] == curr_uf]
    curr_geocode_list = curr_geo_df["geocode"]
    for curr_geocode in curr_geocode_list: 
        result = mosqlient.get_climate(
            api_key = api_key,
            start_date = start_date,
            end_date = end_date,
            uf = curr_uf,
            geocode = curr_geocode
        )
        builder.append(result)

dataset = pd.concat(builder)

dataset.to_csv("output.csv")

dataset
